scripts/requirements.txt
openai>=1.50.0
pyyaml>=6.0
